cam gata proiectu, mai trebuie adaugat captcha la login

+++ src/Controller/LoginController.php
@@ -3,9 +3,12 @@
 namespace App\Controller;
 
 use App\Entity\Account;
+use App\Entity\Comentarii;
 use App\Entity\Stiri;
 use App\Repository\AccountRepository;
+use App\Repository\ComentariiRepository;
 use App\Repository\StiriRepository;
+use CMEN\GoogleChartsBundle\GoogleCharts\Charts\PieChart;
 use http\Client\Request;
 use PhpOffice\PhpSpreadsheet\IOFactory;
 use PhpOffice\PhpSpreadsheet\Reader\Xlsx;
@@ -181,26 +184,41 @@ class LoginController extends AbstractController
         }
     }
     #[Route('/vizionare/{id}', name: 'app_watch')]
-    public function seeArticle(StiriRepository $stiriRepository,\Symfony\Component\HttpFoundation\Request $request): Response
+    public function seeArticle(ComentariiRepository $comentariiRepository,StiriRepository $stiriRepository,\Symfony\Component\HttpFoundation\Request $request): Response
     {
+        $comentarii = $comentariiRepository->findBy(array("id_anunt"=>$request->get("id")));
         $stire = $stiriRepository -> find($request->get("id"));
         return $this->render("main/stire.html.twig",[
-            "stire" => $stire
+            "stire" => $stire,
+            "id" => $request->get("id"),
+            "comentarii" => $comentarii
         ]);
     }
     #[Route('/update/{id}', name: 'app_article_update')]
     public function updateArticle(StiriRepository $stiriRepository,\Symfony\Component\HttpFoundation\Request $request): Response
     {
-        $stire = $stiriRepository -> find($request->get("id"));
-        return $this->render("main/article.html.twig",[
-            "stire" => $stire
-        ]);
+        try {
+            session_start();
+        } catch (\Exception $exception) {
+        }
+        if (isset($_SESSION["rol"]) and ($_SESSION["rol"] == 1 or $_SESSION["rol"] == 2)) {
+            $stire = $stiriRepository->find($request->get("id"));
+            return $this->render("main/article.html.twig", [
+                "stire" => $stire
+            ]);
+        }
+        return new Response("Nu aveti permisiune");
     }
     #[Route('/delete/{id}', name: 'app_article_delete')]
     public function deleteArticle(StiriRepository $stiriRepository,\Symfony\Component\HttpFoundation\Request $request): Response
-    {
+    {        try {
+        session_start();
+    }catch(\Exception $exception){}
+        if(isset($_SESSION["rol"]) and ($_SESSION["rol"] == 1 or $_SESSION["rol"] == 2)){
         $stiriRepository->remove($stiriRepository->find($request->get("id")));
         return $this->redirectToRoute('app_main');
+        }
+        return new Response("Nu aveti permisiune");
     }
     #[Route('/clasament', name: 'app_clasament')]
     public function teamRanks(){
@@ -263,59 +281,148 @@ class LoginController extends AbstractController
     }
     #[Route('/export', name: 'app_export')]
     public function ExportNews(){
-        $spreadsheet = new Spreadsheet();
-        $sheet = $spreadsheet ->getActiveSheet();
-        $sheet->setTitle("Stiri");
-        $sheet->setCellValue("A1", "id");
-        $sheet->setCellValue("B1", "rezumat");
-        $sheet->setCellValue("C1", "text");
-        $sheet->setCellValue("D1", "autor");
-        $sheet->setCellValue("E1", "Poza1");
-        $sheet->setCellValue("F1", "Poza2");
-        $sheet->setCellValue("G1", "Poza3");
-        $stiri = $this->stiriRepository->findAll();
-        for($i=2;$i<=count($stiri)+1;$i++){
-            $sheet->setCellValue("A".$i, $stiri[$i-2]->getId());
-            $sheet->setCellValue("B".$i, $stiri[$i-2]->getTitlu());
-            $sheet->setCellValue("C".$i, $stiri[$i-2]->getRezumat());
-            $sheet->setCellValue("D".$i, $stiri[$i-2]->getText());
-            $sheet->setCellValue("E".$i, $stiri[$i-2]->getAutor());
-            $sheet->setCellValue("F".$i, $stiri[$i-2]->getPoza1());
-            $sheet->setCellValue("G".$i, $stiri[$i-2]->getPoza2());
-            $sheet->setCellValue("H".$i, $stiri[$i-2]->getPoza3());
-        }
-        $writer = new \PhpOffice\PhpSpreadsheet\Writer\Xlsx($spreadsheet);
-        $fileName = 'stiri.xlsx';
-        $temp_file = tempnam(sys_get_temp_dir(), $fileName);
-        $writer -> save($temp_file);
-        return $this->file($temp_file,$fileName,ResponseHeaderBag::DISPOSITION_INLINE);
+            try {
+                session_start();
+            }catch(\Exception $exception){}
+            if($_SESSION["rol"]>=1) {
+                $spreadsheet = new Spreadsheet();
+                $sheet = $spreadsheet->getActiveSheet();
+                $sheet->setTitle("Stiri");
+                $sheet->setCellValue("A1", "id");
+                $sheet->setCellValue("B1", "rezumat");
+                $sheet->setCellValue("C1", "text");
+                $sheet->setCellValue("D1", "autor");
+                $sheet->setCellValue("E1", "Poza1");
+                $sheet->setCellValue("F1", "Poza2");
+                $sheet->setCellValue("G1", "Poza3");
+                $stiri = $this->stiriRepository->findAll();
+                for ($i = 2; $i <= count($stiri) + 1; $i++) {
+                    $sheet->setCellValue("A" . $i, $stiri[$i - 2]->getId());
+                    $sheet->setCellValue("B" . $i, $stiri[$i - 2]->getTitlu());
+                    $sheet->setCellValue("C" . $i, $stiri[$i - 2]->getRezumat());
+                    $sheet->setCellValue("D" . $i, $stiri[$i - 2]->getText());
+                    $sheet->setCellValue("E" . $i, $stiri[$i - 2]->getAutor());
+                    $sheet->setCellValue("F" . $i, $stiri[$i - 2]->getPoza1());
+                    $sheet->setCellValue("G" . $i, $stiri[$i - 2]->getPoza2());
+                    $sheet->setCellValue("H" . $i, $stiri[$i - 2]->getPoza3());
+                }
+                $writer = new \PhpOffice\PhpSpreadsheet\Writer\Xlsx($spreadsheet);
+                $fileName = 'stiri.xlsx';
+                $temp_file = tempnam(sys_get_temp_dir(), $fileName);
+                $writer->save($temp_file);
+                return $this->file($temp_file, $fileName, ResponseHeaderBag::DISPOSITION_INLINE);
+            }
     }
     #[Route('/import', name: 'app_import')]
     public function ImportNews(\Symfony\Component\HttpFoundation\Request $request)
     {
-        $the_file = $request->files->get('uploaded_file');
-        try{
-            $spreadsheet = IOFactory::load($the_file->getRealPath());
-            $sheet        = $spreadsheet->getActiveSheet();
-            $row_limit    = $sheet->getHighestDataRow();
-            $row_range    = range( 2, $row_limit );
-            $data = array();
-            foreach ( $row_range as $row ) {
-                $stire = new Stiri();
-                $stire->setTitlu($sheet->getCell( 'B' . $row )->getValue());
-                $stire->setRezumat($sheet->getCell( 'C' . $row )->getValue());
-                $stire->setText($sheet->getCell( 'D' . $row )->getValue());
-                $stire->setAutor($sheet->getCell( 'E' . $row )->getValue());
-                $stire->setPoza1($sheet->getCell( 'F' . $row )->getValue());
-                $stire->setPoza2($sheet->getCell( 'G' . $row )->getValue());
-                $stire->setPoza3($sheet->getCell( 'H' . $row )->getValue());
-                var_dump($stire);
-                $this->stiriRepository->save($stire);
+        try {
+            session_start();
+        }catch(\Exception $exception){}
+        if($_SESSION["rol"]>=1) {
+            $the_file = $request->files->get('uploaded_file');
+            try {
+                $spreadsheet = IOFactory::load($the_file->getRealPath());
+                $sheet = $spreadsheet->getActiveSheet();
+                $row_limit = $sheet->getHighestDataRow();
+                $row_range = range(2, $row_limit);
+                $data = array();
+                foreach ($row_range as $row) {
+                    $stire = new Stiri();
+                    $stire->setTitlu($sheet->getCell('B' . $row)->getValue());
+                    $stire->setRezumat($sheet->getCell('C' . $row)->getValue());
+                    $stire->setText($sheet->getCell('D' . $row)->getValue());
+                    $stire->setAutor($sheet->getCell('E' . $row)->getValue());
+                    $stire->setPoza1($sheet->getCell('F' . $row)->getValue());
+                    $stire->setPoza2($sheet->getCell('G' . $row)->getValue());
+                    $stire->setPoza3($sheet->getCell('H' . $row)->getValue());
+                    var_dump($stire);
+                    $this->stiriRepository->save($stire);
+                }
+            } catch (Exception $e) {
+                var_dump($e);
             }
-        } catch (Exception $e) {
-            var_dump($e);
+            return new Response("importate");
+        }            return new Response("neimportate");
+    }
+    #[Route('/comenteaza', name: 'app_comentariu')]
+    public function Comentarii(ComentariiRepository $comentariiRepository, StiriRepository $stiriRepository)
+    {
+        try {
+            session_start();
+        } catch (\Exception $exception) {
+        }
+        if(isset($_SESSION["nume"]) and isset($_SESSION["rol"])){
+            $comentariu = new Comentarii();
+            $comentariu->setNumeComentator($_SESSION["nume"]);
+            $comentariu->setIdAnunt($_POST["id_anunt"]);
+            $comentariu->setComentariu($_POST["comentariu"]);
+            $comentariiRepository->save($comentariu);
+        }
+        $comentarii = $comentariiRepository->findBy(array("id_anunt"=>$_POST["id_anunt"]));
+        $stire = $stiriRepository -> find($_POST["id_anunt"]);
+        return $this->redirectToRoute("app_watch",[
+            "stire" => $stire,
+            "id" => $_POST["id_anunt"],
+            "comentarii" => $comentarii
+        ]);
+    }
+    #[Route('/statistici', name: 'app_stats')]
+    public function Stats(ComentariiRepository $comentariiRepository, StiriRepository $stiriRepository){
+        $autoriStiri = [];
+        foreach($stiriRepository->findAll() as $bucata){
+            if(!array_key_exists($bucata->getAutor(),$autoriStiri)){
+                $autoriStiri[$bucata->getAutor()] = 1;
+            }else{
+                $autoriStiri[$bucata->getAutor()]++;
+            }
+        }
+        $autoriComentarii = [];
+        foreach($comentariiRepository->findAll() as $bucata){
+            if(!array_key_exists($bucata->getNumeComentator(),$autoriComentarii)){
+                $autoriComentarii[$bucata->getNumeComentator()] = 1;
+            }else{
+                $autoriComentarii[$bucata->getNumeComentator()]++;
+            }
         }
 
-        return new Response("importate");
+        $stats = array(['Categories','Pie expense']);
+        foreach ($autoriStiri as $key=>$value){
+            array_push($stats,[$key,$value]);
+        }
+        $pieChart = new PieChart();
+        $pieChart->getData()->setArrayToDataTable(
+            $stats
+        );
+        $pieChart->getOptions()->setTitle('Statistici autori stiri');
+        $pieChart->getOptions()->setHeight(200);
+        $pieChart->getOptions()->setWidth(500);
+        $pieChart->getOptions()->getTitleTextStyle()->setBold(true);
+        $pieChart->getOptions()->getTitleTextStyle()->setColor('#009900');
+        $pieChart->getOptions()->getTitleTextStyle()->setItalic(true);
+        $pieChart->getOptions()->getTitleTextStyle()->setFontName('Arial');
+        $pieChart->getOptions()->getTitleTextStyle()->setFontSize(20);
+
+        $stats2 = array(['Categories','Pie expense']);
+        foreach ($autoriComentarii as $key=>$value){
+            array_push($stats2,[$key,$value]);
+        }
+
+        $pieChart2 = new PieChart();
+        $pieChart2->getData()->setArrayToDataTable(
+            $stats2
+        );
+        $pieChart2->getOptions()->setTitle('Statistici autori comentarii');
+        $pieChart2->getOptions()->setHeight(200);
+        $pieChart2->getOptions()->setWidth(500);
+        $pieChart2->getOptions()->getTitleTextStyle()->setBold(true);
+        $pieChart2->getOptions()->getTitleTextStyle()->setColor('#009900');
+        $pieChart2->getOptions()->getTitleTextStyle()->setItalic(true);
+        $pieChart2->getOptions()->getTitleTextStyle()->setFontName('Arial');
+        $pieChart2->getOptions()->getTitleTextStyle()->setFontSize(20);
+        return $this->render("main/stats.html.twig",[
+            "piechart" => $pieChart,
+            "piechart2" => $pieChart2
+        ]);
     }
 }
